Rolling back on the 8-make_multiplier.py

+++ 0x00-python_variable_annotations/8-make_multiplier.py
@@ -5,4 +5,4 @@ from typing import Callable
 
 def make_multiplier(multiplier: float) -> Callable[[float], float]:
     ''' Returns a function that multiplies a float by multiplier '''
-    return float(lambda x: x * multiplier)
+    return lambda x: x * multiplier
